train.py 간단화 시키기 Metirc.py : train.py에서 변경시키지 않을 것 같은 메서드만 따로 모아뒀음
